Optimize code.
