+++ server/service/concurrent_test.go
@@ -38,7 +38,7 @@ func init() {
 func TestServiceController_CreateDependenciesForMircServices(t *testing.T) {
 	tryTimes := 3
 	testCount := 10
-	serviceResource, _, _ := service.AssembleResources(nil)
+	serviceResource, _, _ := service.AssembleResources()
 	for i := 0; i < testCount; i++ {
 		_, err := serviceResource.Create(getContext(), &pb.CreateServiceRequest{
 			Service: &pb.MicroService{
@@ -60,7 +60,7 @@ func TestServiceController_CreateDependenciesForMircServices(t *testing.T) {
 			serviceName := fmt.Sprintf("service%d", i)
 			_, err := serviceResource.CreateDependenciesForMircServices(getContext(), &pb.CreateDependenciesRequest{
 				Dependencies: []*pb.MircroServiceDependency{
-					&pb.MircroServiceDependency{
+					{
 						Consumer: &pb.DependencyMircroService{
 							AppId:       "test_deps",
 							ServiceName: serviceName,
@@ -101,7 +101,7 @@ func TestServiceController_CreateDependenciesForMircServices(t *testing.T) {
 			util.LOGGER.Warnf(nil, "%s: 0.", key)
 			continue
 		}
-		d := &service.MircroServiceDependency{}
+		d := &pb.MicroServiceDependency{}
 		err = json.Unmarshal(resp.Kvs[0].Value, d)
 		if err != nil {
 			util.LOGGER.Errorf(err, "%s failed.", key)
