+++ server/service/microservices.go
@@ -20,10 +20,12 @@ import (
 	"github.com/ServiceComb/service-center/server/core/mux"
 	pb "github.com/ServiceComb/service-center/server/core/proto"
 	"github.com/ServiceComb/service-center/server/core/registry"
+	"github.com/ServiceComb/service-center/server/core/registry/store"
 	"github.com/ServiceComb/service-center/server/infra/quota"
 	"github.com/ServiceComb/service-center/server/plugins/dynamic"
 	"github.com/ServiceComb/service-center/server/service/dependency"
 	ms "github.com/ServiceComb/service-center/server/service/microservice"
+	serviceUtil "github.com/ServiceComb/service-center/server/service/util"
 	"github.com/ServiceComb/service-center/util"
 	"github.com/astaxie/beego"
 	"golang.org/x/net/context"
@@ -165,7 +167,7 @@ func (s *ServiceController) Create(ctx context.Context, in *pb.CreateServiceRequ
 	lock.Unlock()
 	//创建服务间的依赖
 	util.LOGGER.Infof("create microservice: add dependency for %s.operator:%s", serviceFlag, remoteIP)
-	err = s.updateAsProviderDependency(ctx, serviceId, consumer)
+	err = dependency.UpdateAsProviderDependency(ctx, serviceId, consumer)
 	if err != nil {
 		util.LOGGER.Errorf(err, "create microservice failed: dependency update,as provider failed.%s.operator:%s", serviceFlag, remoteIP)
 		return &pb.CreateServiceResponse{
@@ -200,7 +202,7 @@ func (s *ServiceController) Delete(ctx context.Context, in *pb.DeleteServiceRequ
 	force := in.Force
 	tenant := util.ParaseTenantProject(ctx)
 
-	service, err := getServiceByServiceId(ctx, tenant, in.ServiceId)
+	service, err := ms.GetServiceByServiceId(ctx, tenant, in.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "delete microservice failed, serviceId is %s: get service failed.", in.ServiceId)
 		return &pb.DeleteServiceResponse{
@@ -220,7 +222,7 @@ func (s *ServiceController) Delete(ctx context.Context, in *pb.DeleteServiceRequ
 	// 强制删除，则与该服务相关的信息删除，非强制删除： 如果作为该被依赖（作为provider，提供服务,且不是只存在自依赖）或者存在实例，则不能删除
 	if !force {
 		keyConDependency := apt.GenerateProviderDependencyKey(tenant, in.ServiceId, "")
-		services, err := GetDependencies(ctx, keyConDependency, tenant)
+		services, err := dependency.GetDependencies(ctx, keyConDependency, tenant)
 		if err != nil {
 			util.LOGGER.Errorf(err, "delete microservice failed, serviceId is %s:(unforce) inner err, get service dependency failed.", in.ServiceId)
 			return &pb.DeleteServiceResponse{
@@ -265,7 +267,7 @@ func (s *ServiceController) Delete(ctx context.Context, in *pb.DeleteServiceRequ
 	}
 
 	//refresh msCache consumerCache, ensure that watch can notify consumers when no cache.
-	err = refreshDependencyCache(tenant, in.ServiceId, service)
+	err = dependency.RefreshDependencyCache(tenant, in.ServiceId, service)
 	if err != nil {
 		util.LOGGER.Errorf(err, "delete microservice failed, serviceId is %s: inner err, refresh service dependency cache failed.", in.ServiceId)
 		return &pb.DeleteServiceResponse{
@@ -304,7 +306,7 @@ func (s *ServiceController) Delete(ctx context.Context, in *pb.DeleteServiceRequ
 			Response: pb.CreateResponse(pb.Response_FAIL, err.Error()),
 		}, err
 	}
-	optsTmp, err := s.deteleDependencyForService(ctx, consumer, in.ServiceId)
+	optsTmp, err := dependency.DeleteDependencyForService(ctx, consumer, in.ServiceId)
 	lock.Unlock()
 	if err != nil {
 		util.LOGGER.Errorf(err, "delete microservice failed, serviceId is %s: inner err, delete dependency failed.", in.ServiceId)
@@ -347,7 +349,7 @@ func (s *ServiceController) Delete(ctx context.Context, in *pb.DeleteServiceRequ
 	opts = append(opts, opt)
 
 	//删除实例
-	err = deleteServiceAllInstances(ctx, in)
+	err = serviceUtil.DeleteServiceAllInstances(ctx, in)
 	if err != nil {
 		util.LOGGER.Errorf(err, "delete microservice failed, serviceId is %s: delete all instances failed.", in.ServiceId)
 		return &pb.DeleteServiceResponse{
@@ -369,196 +371,6 @@ func (s *ServiceController) Delete(ctx context.Context, in *pb.DeleteServiceRequ
 	}, nil
 }
 
-func refreshDependencyCache(tenant string, providerId string, provider *pb.MicroService) error {
-	key := apt.GenerateProviderDependencyKey(tenant, providerId, "")
-	resp, err := registry.GetRegisterCenter().Do(context.TODO(), &registry.PluginOp{
-		Action:     registry.GET,
-		Key:        []byte(key),
-		WithPrefix: true,
-		KeyOnly:    true,
-	})
-	if err != nil {
-		util.LOGGER.Errorf(err, "refresh dependency cache failed: query service consumers failed, provider id %s", providerId)
-		return err
-	}
-	if resp == nil || len(resp.Kvs) == 0 {
-		util.LOGGER.Infof("refresh dependency cache: this services %s has no consumer dependency.", providerId)
-	} else {
-		ms.MsCache().Set(providerId, provider, 5*time.Minute)
-		dependency.ConsumerCache().Set(providerId, resp.Kvs, 5*time.Minute)
-	}
-	return nil
-}
-
-func deleteServiceAllInstances(ctx context.Context, in *pb.DeleteServiceRequest) error {
-	tenant := util.ParaseTenantProject(ctx)
-
-	instanceLeaseKey := apt.GenerateInstanceLeaseKey(tenant, in.ServiceId, "")
-
-	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action:     registry.GET,
-		Key:        []byte(instanceLeaseKey),
-		WithPrefix: true,
-	})
-	if err != nil {
-		util.LOGGER.Errorf(err, "delete service all instance failed: get instance lease failed.")
-		return err
-	}
-	if resp.Count <= 0 {
-		util.LOGGER.Warnf(nil, "No instances to revoke.")
-		return nil
-	}
-	for _, v := range resp.Kvs {
-		leaseID, _ := strconv.ParseInt(string(v.Value), 10, 64)
-		registry.GetRegisterCenter().LeaseRevoke(ctx, leaseID)
-	}
-	return nil
-}
-
-func (s *ServiceController) deteleDependencyForService(ctx context.Context, consumer *pb.MicroServiceKey, serviceId string) ([]*registry.PluginOp, error) {
-	opts := []*registry.PluginOp{}
-	optsTmps := []*registry.PluginOp{}
-	tenant := consumer.Tenant
-	flag := map[string]bool{}
-	//删除依赖规则
-	conKey := apt.GenerateConsumerDependencyRuleKey(tenant, consumer)
-	err, providerValue := transferToMircroServiceDependency(ctx, conKey)
-	if err != nil {
-		return nil, err
-	}
-	if providerValue != nil && len(providerValue.Dependency) != 0 {
-		proProkey := ""
-		for _, providerRule := range providerValue.Dependency {
-			proProkey = apt.GenerateProviderDependencyRuleKey(tenant, providerRule)
-			err, consumers := transferToMircroServiceDependency(ctx, proProkey)
-			if err != nil {
-				return nil, err
-			}
-			err = deleteDependencyRuleUtil(ctx, consumers, consumer, proProkey)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		opt := &registry.PluginOp{
-			Action: registry.DELETE,
-			Key:    []byte(conKey),
-		}
-		util.LOGGER.Debugf("conKey is %s.", conKey)
-		opts = append(opts, opt)
-	}
-	//作为provider的依赖规则
-	providerKey := apt.GenerateProviderDependencyRuleKey(tenant, consumer)
-	opt := &registry.PluginOp{
-		Action: registry.DELETE,
-		Key:    []byte(providerKey),
-	}
-	util.LOGGER.Debugf("providerKey is %s", providerKey)
-	opts = append(opts, opt)
-
-	//删除依赖关系
-	optsTmps, err = s.deleteDependencyUtil(ctx, "c", tenant, serviceId, flag)
-	if err != nil {
-		return nil, err
-	}
-	opts = append(opts, optsTmps...)
-	util.LOGGER.Debugf("flag is %s", flag)
-	optsTmps, err = s.deleteDependencyUtil(ctx, "p", tenant, serviceId, flag)
-	if err != nil {
-		return nil, err
-	}
-	util.LOGGER.Debugf("flag is %s", flag)
-	opts = append(opts, optsTmps...)
-	return opts, nil
-}
-
-func deleteDependencyRuleUtil(ctx context.Context, mircroServiceDependency *MircroServiceDependency, service *pb.MicroServiceKey, serviceKey string) error {
-	for key, serviceTmp := range mircroServiceDependency.Dependency {
-		if ok := equalServiceDependency(serviceTmp, service); ok {
-			mircroServiceDependency.Dependency = append(mircroServiceDependency.Dependency[:key], mircroServiceDependency.Dependency[key+1:]...)
-			util.LOGGER.Debugf("delete versionRule from %s", serviceTmp.ServiceName)
-			break
-		}
-	}
-	opt := &registry.PluginOp{}
-	if len(mircroServiceDependency.Dependency) == 0 {
-		opt = &registry.PluginOp{
-			Action: registry.DELETE,
-			Key:    []byte(serviceKey),
-		}
-		util.LOGGER.Debugf("serviceKey is .", serviceKey)
-		util.LOGGER.Debugf("After deleting versionRule from %s,provider's consumer is empty.", serviceKey)
-
-	} else {
-		data, err := json.Marshal(mircroServiceDependency)
-		if err != nil {
-			util.LOGGER.Errorf(nil, "Marshal tmpValue failed.")
-			return err
-		}
-		opt = &registry.PluginOp{
-			Action: registry.PUT,
-			Key:    []byte(serviceKey),
-			Value:  []byte(data),
-		}
-		util.LOGGER.Debugf("serviceKey is %s.", serviceKey)
-	}
-	_, err := registry.GetRegisterCenter().Do(ctx, opt)
-	if err != nil {
-		util.LOGGER.Errorf(err, "Submit update dependency failed.")
-		return err
-	}
-	return nil
-}
-
-func (s *ServiceController) deleteDependencyUtil(ctx context.Context, serviceType string, tenant string, serviceId string, flag map[string]bool) ([]*registry.PluginOp, error) {
-	serviceKey := apt.GenerateServiceDependencyKey(serviceType, tenant, serviceId, "")
-	opt := &registry.PluginOp{
-		Action:     registry.GET,
-		Key:        []byte(serviceKey),
-		WithPrefix: true,
-	}
-	rsp, err := registry.GetRegisterCenter().Do(ctx, opt)
-	if err != nil {
-		return nil, err
-	}
-	opts := []*registry.PluginOp{}
-	if rsp != nil {
-		serviceTmpId := ""
-		serviceTmpKey := ""
-		deleteKey := ""
-		for _, kv := range rsp.Kvs {
-			tmpKeyArr := strings.Split(string(kv.Key), "/")
-			serviceTmpId = tmpKeyArr[len(tmpKeyArr)-1]
-			if serviceType == "p" {
-				serviceTmpKey = apt.GenerateConsumerDependencyKey(tenant, serviceTmpId, serviceId)
-				deleteKey = strings.Join([]string{"c", serviceTmpId, serviceId}, "/")
-			} else {
-				serviceTmpKey = apt.GenerateProviderDependencyKey(tenant, serviceTmpId, serviceId)
-				deleteKey = strings.Join([]string{"p", serviceTmpId, serviceId}, "/")
-			}
-			if _, ok := flag[serviceTmpKey]; ok {
-				util.LOGGER.Debugf("serviceTmpKey is more exist.%s", serviceTmpKey)
-				continue
-			}
-			flag[serviceTmpKey] = true
-			util.LOGGER.Infof("delete dependency %s", deleteKey)
-			opt = &registry.PluginOp{
-				Action: registry.DELETE,
-				Key:    []byte(serviceTmpKey),
-			}
-			opts = append(opts, opt)
-		}
-		opt = &registry.PluginOp{
-			Action:     registry.DELETE,
-			Key:        []byte(serviceKey),
-			WithPrefix: true,
-		}
-		util.LOGGER.Infof("delete dependency serviceKey is %s", serviceType+"/"+serviceId)
-		opts = append(opts, opt)
-	}
-	return opts, nil
-}
-
 func (s *ServiceController) GetOne(ctx context.Context, in *pb.GetServiceRequest) (*pb.GetServiceResponse, error) {
 	if in == nil || len(in.ServiceId) == 0 {
 		return &pb.GetServiceResponse{
@@ -574,7 +386,7 @@ func (s *ServiceController) GetOne(ctx context.Context, in *pb.GetServiceRequest
 		}, nil
 	}
 	tenant := util.ParaseTenantProject(ctx)
-	service, err := getServiceByServiceId(ctx, tenant, in.ServiceId)
+	service, err := ms.GetServiceByServiceId(ctx, tenant, in.ServiceId)
 
 	if err != nil {
 		util.LOGGER.Errorf(err, "get microservice failed, serviceId is %s: inner err,get service failed.", in.ServiceId)
@@ -601,7 +413,7 @@ func (s *ServiceController) GetServices(ctx context.Context, in *pb.GetServicesR
 			Response: pb.CreateResponse(pb.Response_FAIL, "Request format invalid"),
 		}, nil
 	}
-	services, err := GetAllServiceUtil(ctx)
+	services, err := ms.GetAllServiceUtil(ctx)
 	if err != nil {
 		util.LOGGER.Errorf(err, "get services failed: inner err.")
 		return &pb.GetServicesResponse{
@@ -633,7 +445,7 @@ func (s *ServiceController) UpdateProperties(ctx context.Context, in *pb.UpdateS
 	tenant := util.ParaseTenantProject(ctx)
 
 	key := apt.GenerateServiceKey(tenant, in.ServiceId)
-	service, err := getServiceByServiceId(ctx, tenant, in.ServiceId)
+	service, err := ms.GetServiceByServiceId(ctx, tenant, in.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "update service properties failed, serviceId is %s: query service failed.", in.ServiceId)
 		return &pb.UpdateServicePropsResponse{
@@ -678,18 +490,6 @@ func (s *ServiceController) UpdateProperties(ctx context.Context, in *pb.UpdateS
 	}, nil
 }
 
-func (s *ServiceController) RuleExist(ctx context.Context, tenant string, serviceId string, attr string, pattern string) bool {
-	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action:    registry.GET,
-		Key:       []byte(apt.GenerateRuleIndexKey(tenant, serviceId, attr, pattern)),
-		CountOnly: true,
-	})
-	if err != nil || resp.Count == 0 {
-		return false
-	}
-	return true
-}
-
 func (s *ServiceController) AddRule(ctx context.Context, in *pb.AddServiceRulesRequest) (*pb.AddServiceRulesResponse, error) {
 	if in == nil || len(in.ServiceId) == 0 || len(in.GetRules()) == 0 {
 		util.LOGGER.Errorf(nil, "add rule failed: invalid parameters.")
@@ -709,7 +509,7 @@ func (s *ServiceController) AddRule(ctx context.Context, in *pb.AddServiceRulesR
 	}
 
 	opts := []*registry.PluginOp{}
-	ruleType, _, err := getServiceRuleType(ctx, tenant, in.ServiceId)
+	ruleType, _, err := serviceUtil.GetServiceRuleType(ctx, tenant, in.ServiceId)
 	util.LOGGER.Debugf("ruleType is %s", ruleType)
 	if err != nil {
 		return &pb.AddServiceRulesResponse{
@@ -738,7 +538,7 @@ func (s *ServiceController) AddRule(ctx context.Context, in *pb.AddServiceRulesR
 		}
 
 		//同一服务，attribute和pattern确定一个rule
-		if s.RuleExist(ctx, tenant, in.ServiceId, rule.Attribute, rule.Pattern) {
+		if serviceUtil.RuleExist(ctx, tenant, in.ServiceId, rule.Attribute, rule.Pattern) {
 			util.LOGGER.Infof("This rule more exists, %s ", in.ServiceId)
 			continue
 		}
@@ -800,28 +600,6 @@ func (s *ServiceController) AddRule(ctx context.Context, in *pb.AddServiceRulesR
 	}, nil
 }
 
-func getServiceRuleType(ctx context.Context, tenant string, serviceId string) (string, int, error) {
-	key := apt.GenerateServiceRuleKey(tenant, serviceId, "")
-	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action:     registry.GET,
-		Key:        []byte(key),
-		WithPrefix: true,
-	})
-	if err != nil {
-		util.LOGGER.Errorf(err, "Get rule failed.%s", err.Error())
-		return "", 0, err
-	}
-	if len(resp.Kvs) == 0 {
-		return "", 0, nil
-	}
-	rule := &pb.ServiceRule{}
-	err = json.Unmarshal(resp.Kvs[0].Value, rule)
-	if err != nil {
-		util.LOGGER.Errorf(err, "Unmarshal rule data failed.%s", err.Error())
-	}
-	return rule.RuleType, len(resp.Kvs), nil
-}
-
 func (s *ServiceController) UpdateRule(ctx context.Context, in *pb.UpdateServiceRuleRequest) (*pb.UpdateServiceRuleResponse, error) {
 	if in == nil || in.GetRule() == nil || len(in.ServiceId) == 0 || len(in.RuleId) == 0 {
 		util.LOGGER.Errorf(nil, "update rule failed: invalid parameters.")
@@ -848,7 +626,7 @@ func (s *ServiceController) UpdateRule(ctx context.Context, in *pb.UpdateService
 	}
 
 	//是否能改变ruleType
-	ruleType, ruleNum, err := getServiceRuleType(ctx, tenant, in.ServiceId)
+	ruleType, ruleNum, err := serviceUtil.GetServiceRuleType(ctx, tenant, in.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "update rule failed, serviceId is %s, ruleId is %s: get rule type failed.", in.ServiceId, in.RuleId)
 		return &pb.UpdateServiceRuleResponse{
@@ -862,30 +640,11 @@ func (s *ServiceController) UpdateRule(ctx context.Context, in *pb.UpdateService
 		}, nil
 	}
 
-	key := apt.GenerateServiceRuleKey(tenant, in.ServiceId, in.RuleId)
-	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action: registry.GET,
-		Key:    []byte(key),
-	})
+	rule, err := serviceUtil.GetOneRule(ctx, tenant, in.ServiceId, in.RuleId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "update rule failed, serviceId is %s, ruleId is %s: query service rule failed.", in.ServiceId, in.RuleId)
 		return &pb.UpdateServiceRuleResponse{
 			Response: pb.CreateResponse(pb.Response_FAIL, "get service rule file failed"),
-		}, err
-	}
-
-	if len(resp.Kvs) == 0 {
-		util.LOGGER.Errorf(err, "update rule failed, serviceId is %s, ruleId is %s:this rule does not exist,can't update.", in.ServiceId, in.RuleId)
-		return &pb.UpdateServiceRuleResponse{
-			Response: pb.CreateResponse(pb.Response_FAIL, "This rule does not exist."),
-		}, err
-	}
-	rule := &pb.ServiceRule{}
-	err = json.Unmarshal(resp.Kvs[0].Value, rule)
-	if err != nil {
-		util.LOGGER.Errorf(err, "update rule failed, serviceId is %s, ruleId is %s: unmarshal service rule failed.", in.ServiceId, in.RuleId)
-		return &pb.UpdateServiceRuleResponse{
-			Response: pb.CreateResponse(pb.Response_FAIL, "unmarshal service rule file failed"),
 		}, err
 	}
 
@@ -904,6 +663,7 @@ func (s *ServiceController) UpdateRule(ctx context.Context, in *pb.UpdateService
 	rule.Description = in.GetRule().Description
 	rule.Timestamp = strconv.FormatInt(time.Now().Unix(), 10)
 
+	key := apt.GenerateServiceRuleKey(tenant, in.ServiceId, in.RuleId)
 	util.LOGGER.Debugf("start update service rule file: %s", key)
 	data, err := json.Marshal(rule)
 	if err != nil {
@@ -946,7 +706,7 @@ func (s *ServiceController) UpdateRule(ctx context.Context, in *pb.UpdateService
 		}, err
 	}
 
-	util.LOGGER.Infof("update rule successful: servieId is , ruleId is %s.", in.ServiceId, in.RuleId)
+	util.LOGGER.Infof("update rule successful: servieId is %s, ruleId is %s.", in.ServiceId, in.RuleId)
 	return &pb.UpdateServiceRuleResponse{
 		Response: pb.CreateResponse(pb.Response_SUCCESS, "get service rules successfully"),
 	}, nil
@@ -969,7 +729,7 @@ func (s *ServiceController) GetRule(ctx context.Context, in *pb.GetServiceRulesR
 		}, nil
 	}
 
-	rules, err := GetRulesUtil(ctx, tenant, in.ServiceId)
+	rules, err := serviceUtil.GetRulesUtil(ctx, tenant, in.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(nil, "get service rule failed, serviceId is %s: get rule failed.", in.ServiceId)
 		return &pb.GetServiceRulesResponse{
@@ -1006,7 +766,7 @@ func (s *ServiceController) DeleteRule(ctx context.Context, in *pb.DeleteService
 	for _, ruleId := range in.RuleIds {
 		key = apt.GenerateServiceRuleKey(tenant, in.ServiceId, ruleId)
 		util.LOGGER.Debugf("start delete service rule file: %s", key)
-		data, err := s.getOneRule(ctx, key, ruleId)
+		data, err := serviceUtil.GetOneRule(ctx, tenant, in.ServiceId, ruleId)
 		if err != nil {
 			util.LOGGER.Errorf(err, "delete service rule failed, serviceId is %s, rule is %v: get rule of ruleId %s failed.", in.ServiceId, in.RuleIds, ruleId)
 			return &pb.DeleteServiceRulesResponse{
@@ -1047,29 +807,6 @@ func (s *ServiceController) DeleteRule(ctx context.Context, in *pb.DeleteService
 	return &pb.DeleteServiceRulesResponse{
 		Response: pb.CreateResponse(pb.Response_SUCCESS, "delete service rules successfully"),
 	}, nil
-}
-
-func (s *ServiceController) getOneRule(ctx context.Context, key string, ruleId string) (*pb.ServiceRule, error) {
-	opt := &registry.PluginOp{
-		Action: registry.GET,
-		Key:    []byte(key),
-	}
-	resp, err := registry.GetRegisterCenter().Do(ctx, opt)
-	if err != nil {
-		util.LOGGER.Errorf(nil, "Get rule for service failed for %s.", err.Error())
-		return nil, err
-	}
-	rule := &pb.ServiceRule{}
-	if len(resp.Kvs) == 0 {
-		util.LOGGER.Errorf(nil, "Get rule failed, ruleId is %s.", ruleId)
-		return nil, nil
-	}
-	err = json.Unmarshal(resp.Kvs[0].Value, rule)
-	if err != nil {
-		util.LOGGER.Errorf(nil, "unmarshal resp failed for %s.", err.Error())
-		return nil, err
-	}
-	return rule, nil
 }
 
 func (s *ServiceController) Exist(ctx context.Context, in *pb.GetExistenceRequest) (*pb.GetExistenceResponse, error) {
@@ -1144,7 +881,7 @@ func (s *ServiceController) Exist(ctx context.Context, in *pb.GetExistenceReques
 		}
 
 		key := apt.GenerateServiceSchemaKey(tenant, in.ServiceId, in.SchemaId)
-		err, exist := s.checkSchemaInfoExist(ctx, key)
+		err, exist := serviceUtil.CheckSchemaInfoExist(ctx, key)
 		if err != nil {
 			util.LOGGER.Errorf(err, "schema exist failed, serviceId %s, schemaId : get schema failed.", in.ServiceId, in.SchemaId)
 			return &pb.GetExistenceResponse{
@@ -1195,41 +932,22 @@ func (s *ServiceController) AddTags(ctx context.Context, in *pb.AddServiceTagsRe
 	}
 
 	addTags := in.GetTags()
-
-	key := apt.GenerateServiceTagKey(tenant, in.ServiceId)
-
-	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action: registry.GET,
-		Key:    []byte(key),
-	})
+	dataTags, err := serviceUtil.GetTagsUtils(ctx, tenant, in.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "add service tags failed, serviceId %s, tags %v: get existed tag failed.", in.ServiceId, in.Tags)
 		return &pb.AddServiceTagsResponse{
 			Response: pb.CreateResponse(pb.Response_FAIL, "Get tags failed."),
 		}, err
 	}
-
-	dataTags := map[string]string{}
-	if len(resp.Kvs) > 0 {
-		err = json.Unmarshal(resp.Kvs[0].Value, &dataTags)
-		if err != nil {
-			util.LOGGER.Errorf(err, "add service tags failed, serviceId %s, tags %v: json unmarshal tag data failed.", in.ServiceId, in.Tags)
-			return &pb.AddServiceTagsResponse{
-				Response: pb.CreateResponse(pb.Response_FAIL, "Unmarshal tags failed."),
-			}, nil
-		}
-		if len(dataTags) > 0 {
-			for key, value := range addTags {
-				dataTags[key] = value
-			}
-		} else {
-			dataTags = addTags
+	if len(dataTags) > 0 {
+		for key, value := range addTags {
+			dataTags[key] = value
 		}
 	} else {
 		dataTags = addTags
 	}
 
-	err = addTagIntoETCD(ctx, tenant, in.ServiceId, dataTags)
+	err = serviceUtil.AddTagIntoETCD(ctx, tenant, in.ServiceId, dataTags)
 	if err != nil {
 		util.LOGGER.Errorf(err, "add service tags failed, serviceId %s, tags %v: commit tag data into etcd failed.", in.ServiceId, in.Tags)
 		return &pb.AddServiceTagsResponse{
@@ -1268,7 +986,7 @@ func (s *ServiceController) UpdateTag(ctx context.Context, in *pb.UpdateServiceT
 		}, nil
 	}
 
-	tags, err := GetTagsUtils(ctx, tenant, in.ServiceId)
+	tags, err := serviceUtil.GetTagsUtils(ctx, tenant, in.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "update service tag failed, serviceId %s, tag %s: get tag failed.", in.ServiceId, tagFlag)
 		return &pb.UpdateServiceTagResponse{
@@ -1284,7 +1002,7 @@ func (s *ServiceController) UpdateTag(ctx context.Context, in *pb.UpdateServiceT
 	}
 	tags[in.Key] = in.Value
 
-	err = addTagIntoETCD(ctx, tenant, in.ServiceId, tags)
+	err = serviceUtil.AddTagIntoETCD(ctx, tenant, in.ServiceId, tags)
 
 	if err != nil {
 		util.LOGGER.Errorf(err, "update service tag failed, serviceId %s, tag %s: adding service tags failed.", in.ServiceId, tagFlag)
@@ -1297,26 +1015,6 @@ func (s *ServiceController) UpdateTag(ctx context.Context, in *pb.UpdateServiceT
 	return &pb.UpdateServiceTagResponse{
 		Response: pb.CreateResponse(pb.Response_SUCCESS, "update service tag success."),
 	}, nil
-}
-
-func addTagIntoETCD(ctx context.Context, tenant string, serviceId string, dataTags map[string]string) error {
-	key := apt.GenerateServiceTagKey(tenant, serviceId)
-	data, err := json.Marshal(dataTags)
-	if err != nil {
-		util.LOGGER.Errorf(err, "add tag into etcd,serviceId %s:json marshal tag data failed.", serviceId)
-		return err
-	}
-
-	_, err = registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action: registry.PUT,
-		Key:    []byte(key),
-		Value:  data,
-	})
-	if err != nil {
-		util.LOGGER.Errorf(err, "add tag into etcd,serviceId %s: commit tag data into etcd failed.", serviceId)
-		return err
-	}
-	return nil
 }
 
 func (s *ServiceController) DeleteTags(ctx context.Context, in *pb.DeleteServiceTagsRequest) (*pb.DeleteServiceTagsResponse, error) {
@@ -1343,29 +1041,11 @@ func (s *ServiceController) DeleteTags(ctx context.Context, in *pb.DeleteService
 		}, nil
 	}
 
-	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action: registry.GET,
-		Key:    []byte(apt.GenerateServiceTagKey(tenant, in.ServiceId)),
-	})
+	tags, err := serviceUtil.GetTagsUtils(ctx, tenant, in.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "delete service tags failed, serviceId %s, tags %v: query service failed.", in.ServiceId, in.Keys)
 		return &pb.DeleteServiceTagsResponse{
 			Response: pb.CreateResponse(pb.Response_FAIL, "get service tags file failed"),
-		}, err
-	}
-	if len(resp.Kvs) == 0 {
-		util.LOGGER.Errorf(err, "delete service tags failed, serviceId %s, tags %v: tag not exist,can't delete.", in.ServiceId, in.Keys)
-		return &pb.DeleteServiceTagsResponse{
-			Response: pb.CreateResponse(pb.Response_FAIL, "tags do not exist"),
-		}, nil
-	}
-
-	var tags map[string]string
-	err = json.Unmarshal(resp.Kvs[0].Value, &tags)
-	if err != nil {
-		util.LOGGER.Errorf(err, "delete service tags failed, serviceId %s, tags %v: unmarshal tag data failed.", in.ServiceId, in.Keys)
-		return &pb.DeleteServiceTagsResponse{
-			Response: pb.CreateResponse(pb.Response_FAIL, "unmarshal tags file failed"),
 		}, err
 	}
 	for _, key := range in.Keys {
@@ -1432,7 +1112,7 @@ func (s *ServiceController) GetTags(ctx context.Context, in *pb.GetServiceTagsRe
 		}, nil
 	}
 
-	tags, err := GetTagsUtils(ctx, tenant, in.ServiceId)
+	tags, err := serviceUtil.GetTagsUtils(ctx, tenant, in.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "get service tags failed, serviceId %s: get tag failed.", in.ServiceId)
 		return &pb.GetServiceTagsResponse{
@@ -1472,7 +1152,7 @@ func (s *ServiceController) GetSchemaInfo(ctx context.Context, request *pb.GetSc
 	}
 
 	key := apt.GenerateServiceSchemaKey(tenant, request.ServiceId, request.SchemaId)
-	resp, errDo := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
+	resp, errDo := store.Store().Schema().Search(ctx, &registry.PluginOp{
 		Action: registry.GET,
 		Key:    []byte(key),
 	})
@@ -1518,7 +1198,7 @@ func (s *ServiceController) DeleteSchema(ctx context.Context, request *pb.Delete
 	}
 
 	key := apt.GenerateServiceSchemaKey(tenant, request.ServiceId, request.SchemaId)
-	err, exist := s.checkSchemaInfoExist(ctx, key)
+	err, exist := serviceUtil.CheckSchemaInfoExist(ctx, key)
 	if err != nil {
 		util.LOGGER.Errorf(err, "delete schema failded, serviceId %s, schemaId %s: get schema failed.", request.ServiceId, request.SchemaId)
 		return &pb.DeleteSchemaResponse{
@@ -1591,7 +1271,7 @@ func (s *ServiceController) canModifySchema(ctx context.Context, request *pb.Mod
 		return err, false
 	}
 	tenant := util.ParaseTenantProject(ctx)
-	service, err := getServiceByServiceId(ctx, tenant, request.ServiceId)
+	service, err := ms.GetServiceByServiceId(ctx, tenant, request.ServiceId)
 	if err != nil {
 		util.LOGGER.Errorf(err, "update schema failded, serviceId %s, schemaId %s: get service failed.", request.ServiceId, request.SchemaId)
 		return err, false
@@ -1619,78 +1299,4 @@ func containsValueInSlice(in []string, value string) bool {
 		}
 	}
 	return false
-}
-
-func (s *ServiceController) checkSchemaInfoExist(ctx context.Context, key string) (error, bool) {
-	resp, errDo := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action:    registry.GET,
-		Key:       []byte(key),
-		CountOnly: true,
-	})
-	if errDo != nil {
-		return errDo, false
-	}
-	if resp.Count == 0 {
-		return nil, false
-	}
-	return nil, true
-}
-
-func GetAllServiceUtil(ctx context.Context) ([]*pb.MicroService, error) {
-	tenant := util.ParaseTenantProject(ctx)
-	services, err := ms.GetServicesByTenant(ctx, tenant)
-	if err != nil {
-		return nil, err
-	}
-	return services, nil
-}
-
-func GetTagsUtils(ctx context.Context, tenant string, serviceId string) (map[string]string, error) {
-	tags := map[string]string{}
-
-	key := apt.GenerateServiceTagKey(tenant, serviceId)
-	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action: registry.GET,
-		Key:    []byte(key),
-	})
-	if err != nil {
-		return tags, err
-	}
-	if len(resp.Kvs) != 0 {
-		util.LOGGER.Debugf("start unmarshal service tags file: %s", key)
-		err = json.Unmarshal(resp.Kvs[0].Value, &tags)
-		if err != nil {
-			return tags, err
-		}
-	}
-	return tags, nil
-}
-
-func GetRulesUtil(ctx context.Context, tenant string, serviceId string) ([]*pb.ServiceRule, error) {
-	key := strings.Join([]string{
-		apt.GetServiceRuleRootKey(tenant),
-		serviceId,
-		"",
-	}, "/")
-
-	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
-		Action:     registry.GET,
-		Key:        []byte(key),
-		WithPrefix: true,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	rules := []*pb.ServiceRule{}
-	for _, kvs := range resp.Kvs {
-		util.LOGGER.Debugf("start unmarshal service rule file: %s", string(kvs.Key))
-		rule := &pb.ServiceRule{}
-		err := json.Unmarshal(kvs.Value, rule)
-		if err != nil {
-			return nil, err
-		}
-		rules = append(rules, rule)
-	}
-	return rules, nil
 }
