+++ server/service/tenant/tenent_service.go
@@ -14,7 +14,7 @@
 package tenant
 
 import (
-	"github.com/ServiceComb/service-center/server/core"
+	apt "github.com/ServiceComb/service-center/server/core"
 	"github.com/ServiceComb/service-center/server/core/registry"
 	"github.com/coreos/etcd/mvcc/mvccpb"
 	"golang.org/x/net/context"
@@ -23,7 +23,7 @@ import (
 
 func GetAllTenantRawData() ([]*mvccpb.KeyValue, error) {
 	opt := &registry.PluginOp{
-		Key:        []byte(core.GenerateTenantKey("")),
+		Key:        []byte(apt.GenerateTenantKey("")),
 		Action:     registry.GET,
 		WithPrefix: true,
 	}
@@ -48,9 +48,21 @@ func GetAllTenent() ([]string, error) {
 		for _, kv := range kvs {
 			arrTmp = strings.Split(string(kv.Key), "/")
 			tenant = arrTmp[len(arrTmp)-1]
-			instByTenant = core.GetInstanceRootKey(tenant)
+			instByTenant = apt.GetInstanceRootKey(tenant)
 			insWatherByTenantKeys = append(insWatherByTenantKeys, instByTenant)
 		}
 	}
 	return insWatherByTenantKeys, err
+}
+
+func NewDomain(ctx context.Context, tenant string) error {
+	opt := &registry.PluginOp{
+		Key:    []byte(apt.GenerateTenantKey(tenant)),
+		Action: registry.PUT,
+	}
+	_, err := registry.GetRegisterCenter().PutNoOverride(ctx, opt)
+	if err != nil {
+		return err
+	}
+	return nil
 }
