+++ server/service/util/rule_util.go
@@ -1,0 +1,98 @@
+package util
+
+import (
+	"encoding/json"
+	apt "github.com/ServiceComb/service-center/server/core"
+	pb "github.com/ServiceComb/service-center/server/core/proto"
+	"github.com/ServiceComb/service-center/server/core/registry"
+	"github.com/ServiceComb/service-center/server/core/registry/store"
+	"github.com/ServiceComb/service-center/util"
+	"golang.org/x/net/context"
+	"strings"
+)
+
+func GetRulesUtil(ctx context.Context, tenant string, serviceId string) ([]*pb.ServiceRule, error) {
+	key := strings.Join([]string{
+		apt.GetServiceRuleRootKey(tenant),
+		serviceId,
+		"",
+	}, "/")
+
+	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
+		Action:     registry.GET,
+		Key:        []byte(key),
+		WithPrefix: true,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	rules := []*pb.ServiceRule{}
+	for _, kvs := range resp.Kvs {
+		util.LOGGER.Debugf("start unmarshal service rule file: %s", string(kvs.Key))
+		rule := &pb.ServiceRule{}
+		err := json.Unmarshal(kvs.Value, rule)
+		if err != nil {
+			return nil, err
+		}
+		rules = append(rules, rule)
+	}
+	return rules, nil
+}
+
+func RuleExist(ctx context.Context, tenant string, serviceId string, attr string, pattern string) bool {
+	resp, err := store.Store().RuleIndex().Search(ctx, &registry.PluginOp{
+		Action:    registry.GET,
+		Key:       []byte(apt.GenerateRuleIndexKey(tenant, serviceId, attr, pattern)),
+		CountOnly: true,
+	})
+	if err != nil || resp.Count == 0 {
+		return false
+	}
+	return true
+}
+
+func GetServiceRuleType(ctx context.Context, tenant string, serviceId string) (string, int, error) {
+	key := apt.GenerateServiceRuleKey(tenant, serviceId, "")
+	resp, err := registry.GetRegisterCenter().Do(ctx, &registry.PluginOp{
+		Action:     registry.GET,
+		Key:        []byte(key),
+		WithPrefix: true,
+	})
+	if err != nil {
+		util.LOGGER.Errorf(err, "Get rule failed.%s", err.Error())
+		return "", 0, err
+	}
+	if len(resp.Kvs) == 0 {
+		return "", 0, nil
+	}
+	rule := &pb.ServiceRule{}
+	err = json.Unmarshal(resp.Kvs[0].Value, rule)
+	if err != nil {
+		util.LOGGER.Errorf(err, "Unmarshal rule data failed.%s", err.Error())
+	}
+	return rule.RuleType, len(resp.Kvs), nil
+}
+
+func GetOneRule(ctx context.Context, tenant, serviceId, ruleId string) (*pb.ServiceRule, error) {
+	opt := &registry.PluginOp{
+		Action: registry.GET,
+		Key:    []byte(apt.GenerateServiceRuleKey(tenant, serviceId, ruleId)),
+	}
+	resp, err := store.Store().Rule().Search(ctx, opt)
+	if err != nil {
+		util.LOGGER.Errorf(nil, "Get rule for service failed for %s.", err.Error())
+		return nil, err
+	}
+	rule := &pb.ServiceRule{}
+	if len(resp.Kvs) == 0 {
+		util.LOGGER.Errorf(nil, "Get rule failed, ruleId is %s.", ruleId)
+		return nil, nil
+	}
+	err = json.Unmarshal(resp.Kvs[0].Value, rule)
+	if err != nil {
+		util.LOGGER.Errorf(nil, "unmarshal resp failed for %s.", err.Error())
+		return nil, err
+	}
+	return rule, nil
+}
